refactor password func into helpers
