+++ smart-contracts/test/Lock/hooks/DiscountCodeHook.js
@@ -2,28 +2,7 @@ const { assert } = require('chai')
 const { expect } = require('chai')
 const { ethers, unlock } = require('hardhat')
 
-/**
- * Helper function
- * @param {*} password
- * @param {*} message
- * @returns
- */
-const getSignatureForPassword = async (password, message) => {
-  // Build the signer
-  const encoded = ethers.utils.defaultAbiCoder.encode(
-    ['bytes32'],
-    [ethers.utils.id(password)]
-  )
-  const privateKey = ethers.utils.keccak256(encoded)
-  const privateKeyAccount = new ethers.Wallet(privateKey)
-
-  // Sign
-  const messageHash = ethers.utils.solidityKeccak256(['string'], [message])
-  const messageHashBinary = ethers.utils.arrayify(messageHash)
-  const signature = await privateKeyAccount.signMessage(messageHashBinary)
-
-  return [signature, privateKeyAccount.address]
-}
+const { getSignatureForPassword } = require('../../helpers')
 
 describe('DiscountHook', function () {
   it('should work', async function () {
