+++ smart-contracts/test/helpers/index.js
@@ -16,6 +16,7 @@ const versions = require('./versions')
 const bridge = require('./bridge')
 const events = require('./events')
 const bigNumber = require('./bigNumber')
+const password = require('./password')
 
 module.exports = {
   deployContracts,
@@ -36,4 +37,5 @@ module.exports = {
   ...bridge,
   ...events,
   ...bigNumber,
+  ...password,
 }
