+++ smart-contracts/test/helpers/password.js
@@ -1,0 +1,28 @@
+const { ethers } = require('hardhat')
+
+/**
+ * Helper function
+ * @param {*} password
+ * @param {*} message
+ * @returns
+ */
+const getSignatureForPassword = async (password, message) => {
+  // Build the signer
+  const encoded = ethers.utils.defaultAbiCoder.encode(
+    ['bytes32'],
+    [ethers.utils.id(password)]
+  )
+  const privateKey = ethers.utils.keccak256(encoded)
+  const privateKeyAccount = new ethers.Wallet(privateKey)
+
+  // Sign
+  const messageHash = ethers.utils.solidityKeccak256(['string'], [message])
+  const messageHashBinary = ethers.utils.arrayify(messageHash)
+  const signature = await privateKeyAccount.signMessage(messageHashBinary)
+
+  return [signature, privateKeyAccount.address]
+}
+
+module.exports = {
+  getSignatureForPassword,
+}
